Feat- turretStats
Can upgrade speed turret
Can upgrade damage turret
Can delete turret
With UI of the turret when you click on turret

+++ Assets/Scripts/Life&Damage/LifeManager.cs
@@ -8,9 +8,9 @@ public class LifeManager : MonoBehaviour
     [SerializeField] float hp = 100f;
     public float maxHp = 100f;
 
-    public int damage;
+    public float damage;
 
-    public void TakeDamage(int _damage)
+    public void TakeDamage(float _damage)
     {
         
         hp -= _damage;
